Main: fix warning (size_t assigned to ULONG)
svn:sync-xref-src-repo-rev: r34005

+++ src/VBox/Main/MachineImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: MachineImpl.cpp 11091 2008-08-04 08:44:56Z noreply@oracle.com $ */
+/* $Id: MachineImpl.cpp 11093 2008-08-04 09:04:55Z noreply@oracle.com $ */
 /** @file
  * Implementation of IMachine in VBoxSVC.
  */
@@ -8900,7 +8900,7 @@ STDMETHODIMP SessionMachine::PullGuestProperties (ComSafeArrayOut(BSTR, aNames),
     AssertReturn(!ComSafeArrayOutIsNull (aTimestamps), E_POINTER);
     AssertReturn(!ComSafeArrayOutIsNull (aFlags), E_POINTER);
 
-    unsigned cEntries = mHWData->mGuestProperties.size();
+    size_t cEntries = mHWData->mGuestProperties.size();
     com::SafeArray <BSTR> names(cEntries);
     com::SafeArray <BSTR> values(cEntries);
     com::SafeArray <ULONG64> timestamps(cEntries);
